.cursor/skills/official-doc-drafter/scripts/requirements.txt
python-docx>=0.8.11
docxtpl>=0.16.0  # python-docx-template: Jinja2 in .docx ({{ var }}, {% for %})
PyMuPDF>=1.22.5
requests>=2.31.0
Pillow>=10.0.0
cairosvg>=2.6.0  # Optional: SVG→PNG
